+++ tests/test_cli.py
@@ -19,15 +19,22 @@ from utilities import check_expected
 
 def clean_images_data(images):
     """
-    Clean an Image.to_dict() for testing
+    Clean an a list of Image.to_dict() for testing
     """
-    for i in images:
-        i['extracted_location'] = None
-        i['archive_location'] = None
-        for l in i['layers']:
-            l['extracted_location'] = None
-            l['archive_location'] = None
+    for image in images:
+        clean_image_data(image)
     return images
+
+
+def clean_image_data(image):
+    """
+    Clean `image` data from Image.to_dict() for testing
+    """
+    image['extracted_location'] = ''
+    image['archive_location'] = os.path.basename(image['archive_location'] or '')
+    
+    return image
+
 
 
 class TestContainerInspectorCli(FileBasedTesting):
@@ -43,14 +50,14 @@ class TestContainerInspectorCli(FileBasedTesting):
     def test_container_inspector_single_layer_from_tarball(self):
         test_dir = self.get_test_loc('cli/hello-world.tar')
         expected = self.get_test_loc('cli/hello-world.tar-inventory-from-tarball-expected.json')
-        out = cli._container_inspector(image_path=test_dir)
+        out = cli._container_inspector(image_path=test_dir, _layer_path_segments=1)
         result = clean_images_data(json.loads(out))
         check_expected(result, expected, regen=False)
 
     def test_container_inspector_multiple_layers_from_tarball(self):
         test_dir = self.get_test_loc('cli/she-image_from_scratch-1.0.tar')
         expected = self.get_test_loc('cli/she-image_from_scratch-1.0.tar-inventory-from-tarball-expected.json')
-        out = cli._container_inspector(image_path=test_dir)
+        out = cli._container_inspector(image_path=test_dir, _layer_path_segments=1)
         result = clean_images_data(json.loads(out))
         check_expected(result, expected, regen=False)